+++ app/build.gradle
@@ -62,4 +62,11 @@ dependencies {
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_ui_version"
     implementation 'com.squareup.retrofit2:retrofit:2.7.1'
     implementation 'com.squareup.retrofit2:converter-gson:2.7.1'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.2.0'
+    androidTestImplementation 'androidx.navigation:navigation-testing:2.3.0'
+    implementation 'androidx.navigation:navigation-compose:2.4.0'
+
+
+
+
 }
