+++ app/src/main/java/com/example/workoutapp/MainActivity.kt
@@ -1,6 +1,7 @@
 package com.example.workoutapp
 
 import android.os.Bundle
+import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
@@ -13,42 +14,32 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.example.workoutapp.ui.theme.WorkoutAppTheme
 
 
-import android.speech.tts.TextToSpeech
-import androidx.activity.compose.setContent
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
-import androidx.compose.material.MaterialTheme.colors
-import androidx.compose.material.MaterialTheme.shapes
-import androidx.compose.material.MaterialTheme.typography
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Home
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Color.Companion.Gray
-import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.clipRect
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Popup
-import com.example.workoutapp.ui.theme.WorkoutAppTheme
+import androidx.navigation.NavHostController
+import androidx.navigation.NavType
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.navArgument
 import kotlinx.coroutines.launch
 import java.text.SimpleDateFormat
 import java.util.*
@@ -60,6 +51,7 @@ class MainActivity : ComponentActivity() {
         setContent {
 
             WorkoutAppTheme {
+
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
@@ -73,8 +65,11 @@ class MainActivity : ComponentActivity() {
     }
 }
 
+@Preview(showBackground = true)
 @Composable
-fun Main(name: String) {
+fun Main(name: String?) {
+
+//Navigation is setup with a Scaffold
     Column(    modifier = Modifier,
         verticalArrangement = Arrangement.spacedBy(10.dp),
     ) {   val scaffoldState = rememberScaffoldState()
@@ -84,18 +79,19 @@ fun Main(name: String) {
             scaffoldState = scaffoldState,
 
             topBar = {
-                AppBar (
-                    onNavigationIconClick = {
-                        scope.launch {
-                            scaffoldState.drawerState.open()
-                        }
 
+                AppBar {
+                    scope.launch {
+                        scaffoldState.drawerState.open()
                     }
-                )
+
+                }
+
             },
             drawerGesturesEnabled = scaffoldState.drawerState.isOpen ,
             drawerContent = {
                 DrawerHeader()
+                // Items within the open drawer
                 DrawerBody(
                     items = listOf(
                         MenuItem(
@@ -125,7 +121,7 @@ fun Main(name: String) {
 content = {
     Column() {
         Text(
-            text = "Hello $name!",
+            text = "Hello !",
             fontSize = 32.sp,
             fontWeight = FontWeight.Bold,
             modifier = Modifier
@@ -246,129 +242,8 @@ content = {
     }
 }
 
-@Composable
-fun Profile(name: String) {
-    Column(
-        modifier = Modifier,
-        verticalArrangement = Arrangement.spacedBy(10.dp),
-    ) {
-        Text(
-            text = "$name's Profile",
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier
-                .padding(20.dp)
-                .width(1000.dp),
-            textAlign = TextAlign.Center
-
-        )
-        Text(
-            text = "Height",
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier
-                .padding(20.dp)
-                .width(1000.dp),
-            textAlign = TextAlign.Center
-
-        )
-        Text(
-            text = "Weight",
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier
-                .padding(20.dp)
-                .width(1000.dp),
-            textAlign = TextAlign.Center
-
-        )
-        Text(
-            text = "BMI",
-            fontSize = 32.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier
-                .padding(20.dp)
-                .width(1000.dp),
-            textAlign = TextAlign.Center
-
-        )
-        Row(modifier = Modifier
-            .fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center) {
-
-            Text(text = "Goals",
-                fontSize = 25.sp,
-                fontWeight = FontWeight.Bold,
-                textDecoration = TextDecoration.Underline,
-                modifier = Modifier
-                    .padding(10.dp)
-                    .width(1000.dp)
-                ,
-                textAlign = TextAlign.Center
-
-            )
-
-            Text(text = "Workouts",
-                fontSize = 25.sp,
-                fontWeight = FontWeight.Bold,
-                textDecoration = TextDecoration.Underline,
-                modifier = Modifier
-                    .padding(10.dp)
-                    .width(1000.dp)
-                ,
-                textAlign = TextAlign.Center
-
-            )
-        }
-        Row(modifier = Modifier
-            .fillMaxWidth(),
-            horizontalArrangement = Arrangement.Center) {
-            Box(
-                modifier = Modifier
-                    .background(MaterialTheme.colors.primary)
-                    .size(200.dp, 250.dp)
-                    .drawBehind {
-                        val borderSize = 4.dp.toPx()
-                        val y = size.height - borderSize / 2
-                        drawLine(
-                            color = Color.DarkGray,
-                            start = Offset(0f, y),
-                            end = Offset(size.width, y),
-                            strokeWidth = borderSize
-                        )
-                    }
-                    .fillMaxWidth()
-                    .clip(shape = RoundedCornerShape(70)),
-
-                )
-
-            Box(
-                modifier = Modifier
-                    .background(MaterialTheme.colors.primary)
-                    .size(200.dp, 250.dp)
-                    .drawBehind {
-                        val borderSize = 4.dp.toPx()
-                        val y = size.height - borderSize / 2
-                        drawLine(
-                            color = Color.DarkGray,
-                            start = Offset(0f, y),
-                            end = Offset(size.width, y),
-                            strokeWidth = borderSize
-                        )
-                    }
-                    .fillMaxWidth()
-                    .clip(shape = RoundedCornerShape(70)),
-
-                )
-
-        }
 
 
-
-    }
-}
-
-@Preview(showBackground = true)
 @Composable
 fun DefaultPreview() {
     WorkoutAppTheme {
