+++ app/src/test/java/com/example/workoutapp/MainPageUnitTest.kt
@@ -1,0 +1,29 @@
+package com.example.workoutapp
+
+import androidx.lifecycle.Observer
+import org.junit.Test
+
+import org.junit.Assert.*
+
+import io.mockk.MockKAnnotations
+import io.mockk.coEvery
+import io.mockk.coVerify
+import io.mockk.every
+import io.mockk.impl.annotations.MockK
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.newSingleThreadContext
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.runTest
+import kotlinx.coroutines.test.setMain
+import org.junit.After
+
+import org.junit.Before
+import org.junit.Rule
+import org.junit.rules.TestRule
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.TimeUnit
+
+class MainPageUnitTest {
+}
