+++ app/src/test/java/com/example/workoutapp/WorkoutUnitTest.kt
@@ -90,6 +90,7 @@ class WorkoutUnitTest {
         assertTrue(workout.ToString().equals("Barbell Curl pull beginner isolation barbell biceps forearms Stand up with your torso upright while holding a barbell at a shoulder-width grip. The palm of your hands should be facing forward and the elbows should be close to the torso. This will be your starting position. While holding the upper arms stationary, curl the weights forward while contracting the biceps as you breathe out. Tip: Only the forearms should move. Continue the movement until your biceps are fully contracted and the bar is at shoulder level. Hold the contracted position for a second and squeeze the biceps hard. Slowly begin to bring the bar back to starting position as your breathe in. Repeat for the recommended amount of repetitions. strength"))
     }
 
+
     @Test
     fun `Given service connects to Countries JSON stream when data are read and parsed then country collection should be greater than zero`() =runTest {
             launch(Dispatchers.Main) {
